Fix Arrow decimal type conversion to float instead of integer

+++ rust/perspective-python/perspective/tests/table/test_table_arrow.py
@@ -185,7 +185,7 @@ class TestTableArrow(object):
         tbl = Table(arrow_data)
         assert tbl.size() == 10
         assert tbl.schema() == {
-            "a": "integer",
+            "a": "float",
         }
         assert tbl.view().to_columns() == {"a": data[0]}
 
@@ -389,7 +389,7 @@ class TestTableArrow(object):
         tbl = Table(arrow_data)
         assert tbl.size() == 10
         assert tbl.schema() == {
-            "a": "integer",
+            "a": "float",
         }
         assert tbl.view().to_columns() == {"a": data[0]}
